Refactor all Xiaomi sensors
