+++ esphome/components/xiaomi_ble/sensor/xiaomi_sensor.h
@@ -8,43 +8,38 @@
 #ifdef ARDUINO_ARCH_ESP32
 
 namespace esphome {
-namespace xiaomi_cgg1 {
+namespace xiaomi_ble {
 
-class XiaomiCGG1 : public Component, public esp32_ble_tracker::ESPBTDeviceListener {
+class XiaomiSensor : public Component, public esp32_ble_tracker::ESPBTDeviceListener {
  public:
   void set_address(uint64_t address) { address_ = address; }
 
-  bool parse_device(const esp32_ble_tracker::ESPBTDevice &device) override {
-    if (device.address_uint64() != this->address_)
-      return false;
-
-    auto res = xiaomi_ble::parse_xiaomi(device);
-    if (!res.has_value())
-      return false;
-
-    if (res->temperature.has_value() && this->temperature_ != nullptr)
-      this->temperature_->publish_state(*res->temperature);
-    if (res->humidity.has_value() && this->humidity_ != nullptr)
-      this->humidity_->publish_state(*res->humidity);
-    if (res->battery_level.has_value() && this->battery_level_ != nullptr)
-      this->battery_level_->publish_state(*res->battery_level);
-    return true;
-  }
+  bool parse_device(const esp32_ble_tracker::ESPBTDevice &device) override;
 
   void dump_config() override;
   float get_setup_priority() const override { return setup_priority::DATA; }
   void set_temperature(sensor::Sensor *temperature) { temperature_ = temperature; }
   void set_humidity(sensor::Sensor *humidity) { humidity_ = humidity; }
   void set_battery_level(sensor::Sensor *battery_level) { battery_level_ = battery_level; }
+  void set_moisture(sensor::Sensor *moisture) { moisture_ = moisture; }
+  void set_conductivity(sensor::Sensor *conductivity) { conductivity_ = conductivity; }
+  void set_illuminance(sensor::Sensor *illuminance) { illuminance_ = illuminance; }
+  void set_impedance(sensor::Sensor *impedance) { impedance_ = impedance; }
+  void set_weight(sensor::Sensor *weight) { weight_ = weight; }
 
  protected:
   uint64_t address_;
   sensor::Sensor *temperature_{nullptr};
   sensor::Sensor *humidity_{nullptr};
   sensor::Sensor *battery_level_{nullptr};
+  sensor::Sensor *moisture_{nullptr};
+  sensor::Sensor *conductivity_{nullptr};
+  sensor::Sensor *illuminance_{nullptr};
+  sensor::Sensor *weight_{nullptr};
+  sensor::Sensor *impedance_{nullptr};
 };
 
-}  // namespace xiaomi_cgg1
+}  // namespace xiaomi_ble
 }  // namespace esphome
 
 #endif
