+++ esphome/components/xiaomi_mijia/sensor.py
@@ -1,3 +1,3 @@
 import esphome.config_validation as cv
 
-CONFIG_SCHEMA = cv.invalid("This sensor has been renamed to xiaomi_lywsdcgq")
+CONFIG_SCHEMA = cv.invalid("This sensor has been renamed to xiaomi_ble")
