--- esphome/components/xiaomi_xmtzc0xhm/__init__.py
@@ -1,1 +1,0 @@
-# this file can't be submited as an empty one.
